pyproject.toml
[tool.poetry]
name = "random-earnings-bot"
version = "0.1.0"
description = "A Telegram bot for cryptocurrency trading and AI-driven market analysis"
authors = ["Beaver <maxymviunov@pm.me>"]
license = "GPL-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
python-telegram-bot = "^21.6"
python-dotenv = "^1.0.1"

[tool.poetry.dev-dependencies]
pytest = "^7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
